+++ relay-server/src/services/buffer/envelope_stack/sqlite.rs
@@ -97,7 +97,9 @@ impl SqliteEnvelopeStack {
 
         // We convert envelopes into a format which simplifies insertion in the store. If an
         // envelope can't be serialized, we will not insert it.
-        let envelopes = envelopes.iter().filter_map(|e| e.as_ref().try_into().ok());
+        let envelopes = relay_statsd::metric!(timer(RelayTimers::BufferEnvelopesSerialization), {
+            envelopes.iter().filter_map(|e| e.as_ref().try_into().ok())
+        });
 
         // When early return here, we are acknowledging that the elements that we popped from
         // the buffer are lost in case of failure. We are doing this on purposes, since if we were