+++ relay-server/src/services/processor.rs
@@ -45,7 +45,7 @@ use smallvec::{smallvec, SmallVec};
 #[cfg(feature = "processing")]
 use {
     crate::services::store::{Store, StoreEnvelope},
-    crate::utils::{sample, Enforcement, EnvelopeLimiter, ItemAction},
+    crate::utils::{CheckLimits, Enforcement, EnvelopeLimiter, ItemAction},
     itertools::Itertools,
     relay_cardinality::{
         CardinalityLimit, CardinalityLimiter, CardinalityLimitsSplit, RedisSetLimiter,
@@ -72,10 +72,10 @@ use crate::services::global_config::GlobalConfigHandle;
 use crate::services::metrics::{Aggregator, MergeBuckets};
 use crate::services::outcome::{DiscardReason, Outcome, TrackOutcome};
 use crate::services::processor::event::FiltersStatus;
-use crate::services::project::{ProjectInfo, ProjectState};
-use crate::services::project_cache::{
+use crate::services::projects::cache::{
     AddMetricMeta, BucketSource, ProcessMetrics, ProjectCache, UpdateRateLimits,
 };
+use crate::services::projects::project::{ProjectInfo, ProjectState};
 use crate::services::test_store::{Capture, TestStore};
 use crate::services::upstream::{
     SendRequest, UpstreamRelay, UpstreamRequest, UpstreamRequestError,
@@ -608,29 +608,37 @@ impl ProcessingExtractedMetrics {
     }
 
     /// Extends the contained project metrics.
-    pub fn extend_project_metrics(
+    pub fn extend_project_metrics<I>(
         &mut self,
-        mut buckets: Vec<Bucket>,
+        buckets: I,
         sampling_decision: Option<SamplingDecision>,
-    ) {
-        for bucket in &mut buckets {
-            bucket.metadata.extracted_from_indexed =
-                sampling_decision == Some(SamplingDecision::Keep);
-        }
-        self.metrics.project_metrics.extend(buckets);
+    ) where
+        I: IntoIterator<Item = Bucket>,
+    {
+        self.metrics
+            .project_metrics
+            .extend(buckets.into_iter().map(|mut bucket| {
+                bucket.metadata.extracted_from_indexed =
+                    sampling_decision == Some(SamplingDecision::Keep);
+                bucket
+            }));
     }
 
     /// Extends the contained sampling metrics.
-    pub fn extend_sampling_metrics(
+    pub fn extend_sampling_metrics<I>(
         &mut self,
-        mut buckets: Vec<Bucket>,
+        buckets: I,
         sampling_decision: Option<SamplingDecision>,
-    ) {
-        for bucket in &mut buckets {
-            bucket.metadata.extracted_from_indexed =
-                sampling_decision == Some(SamplingDecision::Keep);
-        }
-        self.metrics.sampling_metrics.extend(buckets);
+    ) where
+        I: IntoIterator<Item = Bucket>,
+    {
+        self.metrics
+            .sampling_metrics
+            .extend(buckets.into_iter().map(|mut bucket| {
+                bucket.metadata.extracted_from_indexed =
+                    sampling_decision == Some(SamplingDecision::Keep);
+                bucket
+            }));
     }
 
     /// Applies rate limits to the contained metrics.
@@ -638,24 +646,59 @@ impl ProcessingExtractedMetrics {
     /// This is used to apply rate limits which have been enforced on sampled items of an envelope
     /// to also consistently apply to the metrics extracted from these items.
     #[cfg(feature = "processing")]
-    fn apply_enforcement(&mut self, enforcement: &Enforcement) {
-        for (namespace, limit) in [
-            (MetricNamespace::Transactions, &enforcement.event),
-            (MetricNamespace::Spans, &enforcement.spans),
+    fn apply_enforcement(&mut self, enforcement: &Enforcement, enforced_consistently: bool) {
+        // Metric namespaces which need to be dropped.
+        let mut drop_namespaces: SmallVec<[_; 2]> = smallvec![];
+        // Metrics belonging to this metric namespace need to have the `extracted_from_indexed`
+        // flag reset to `false`.
+        let mut reset_extracted_from_indexed: SmallVec<[_; 2]> = smallvec![];
+
+        for (namespace, limit, indexed) in [
+            (
+                MetricNamespace::Transactions,
+                &enforcement.event,
+                &enforcement.event_indexed,
+            ),
+            (
+                MetricNamespace::Spans,
+                &enforcement.spans,
+                &enforcement.spans_indexed,
+            ),
         ] {
             if limit.is_active() {
-                relay_log::trace!(
-                    "dropping {namespace} metrics, due to enforced limit on envelope"
-                );
-                self.retain(|bucket| bucket.name.try_namespace() != Some(namespace));
+                drop_namespaces.push(namespace);
+            } else if indexed.is_active() && !enforced_consistently {
+                // If the enforcment was not computed by consistently checking the limits,
+                // the quota for the metrics has not yet been incremented.
+                // In this case we have a dropped indexed payload but a metric which still needs to
+                // be accounted for, make sure the metric will still be rate limited.
+                reset_extracted_from_indexed.push(namespace);
             }
+        }
+
+        if !drop_namespaces.is_empty() || !reset_extracted_from_indexed.is_empty() {
+            self.retain_mut(|bucket| {
+                let Some(namespace) = bucket.name.try_namespace() else {
+                    return true;
+                };
+
+                if drop_namespaces.contains(&namespace) {
+                    return false;
+                }
+
+                if reset_extracted_from_indexed.contains(&namespace) {
+                    bucket.metadata.extracted_from_indexed = false;
+                }
+
+                true
+            });
         }
     }
 
     #[cfg(feature = "processing")]
-    fn retain(&mut self, mut f: impl FnMut(&Bucket) -> bool) {
-        self.metrics.project_metrics.retain(&mut f);
-        self.metrics.sampling_metrics.retain(&mut f);
+    fn retain_mut(&mut self, mut f: impl FnMut(&mut Bucket) -> bool) {
+        self.metrics.project_metrics.retain_mut(&mut f);
+        self.metrics.sampling_metrics.retain_mut(&mut f);
     }
 }
 
@@ -721,6 +764,10 @@ struct ProcessEnvelopeState<'a, Group> {
 
     /// The state of the project that this envelope belongs to.
     project_info: Arc<ProjectInfo>,
+
+    /// Currently active cached rate limits of the project this envelope belongs to.
+    #[cfg_attr(not(feature = "processing"), expect(dead_code))]
+    rate_limits: RateLimits,
 
     /// The config of this Relay instance.
     config: Arc<Config>,
@@ -832,9 +879,15 @@ pub struct ProcessEnvelopeResponse {
 ///  - Rate limiters and inbound filters on events in processing mode.
 #[derive(Debug)]
 pub struct ProcessEnvelope {
+    /// Envelope to process.
     pub envelope: ManagedEnvelope,
+    /// The project info.
     pub project_info: Arc<ProjectInfo>,
+    /// Currently active cached rate limits for this project.
+    pub rate_limits: RateLimits,
+    /// Root sampling project info.
     pub sampling_project_info: Option<Arc<ProjectInfo>>,
+    /// Sampling reservoir counters.
     pub reservoir_counters: ReservoirCounters,
 }
 
@@ -1251,6 +1304,7 @@ impl EnvelopeProcessorService {
         mut managed_envelope: TypedEnvelope<G>,
         project_id: ProjectId,
         project_info: Arc<ProjectInfo>,
+        rate_limits: RateLimits,
         sampling_project_info: Option<Arc<ProjectInfo>>,
         reservoir_counters: Arc<Mutex<BTreeMap<RuleId, i64>>>,
     ) -> ProcessEnvelopeState<G> {
@@ -1291,6 +1345,7 @@ impl EnvelopeProcessorService {
             metrics: Metrics::default(),
             extracted_metrics,
             project_info,
+            rate_limits,
             config,
             sampling_project_info,
             project_id,
@@ -1305,55 +1360,25 @@ impl EnvelopeProcessorService {
         &self,
         state: &mut ProcessEnvelopeState<G>,
     ) -> Result<(), ProcessingError> {
+        let global_config = self.inner.global_config.current();
         let rate_limiter = match self.inner.rate_limiter.as_ref() {
             Some(rate_limiter) => rate_limiter,
             None => return Ok(()),
         };
 
-        let project_info = &state.project_info;
-        let global_config = self.inner.global_config.current();
-        let quotas = CombinedQuotas::new(&global_config, project_info.get_quotas());
+        // Cached quotas first, they are quick to evaluate and some quotas (indexed) are not
+        // applied in the fast path, all cached quotas can be applied here.
+        let _ = RateLimiter::Cached.enforce(&global_config, state)?;
 
-        if quotas.is_empty() {
-            return Ok(());
-        }
+        // Enforce all quotas consistently with Redis.
+        let limits = RateLimiter::Consistent(rate_limiter).enforce(&global_config, state)?;
 
-        let event_category = state.event_category();
-
-        // When invoking the rate limiter, capture if the event item has been rate limited to also
-        // remove it from the processing state eventually.
-        let mut envelope_limiter = EnvelopeLimiter::new(|item_scope, quantity| {
-            rate_limiter.is_rate_limited(quotas, item_scope, quantity, false)
-        });
-
-        // Tell the envelope limiter about the event, since it has been removed from the Envelope at
-        // this stage in processing.
-        if let Some(category) = event_category {
-            envelope_limiter.assume_event(category);
-        }
-
-        let scoping = state.managed_envelope.scoping();
-        let (enforcement, limits) = metric!(timer(RelayTimers::EventProcessingRateLimiting), {
-            envelope_limiter.compute(state.managed_envelope.envelope_mut(), &scoping)?
-        });
-        let event_active = enforcement.is_event_active();
-
-        // Use the same rate limits as used for the envelope on the metrics.
-        // Those rate limits should not be checked for expiry or similar to ensure a consistent
-        // limiting of envelope items and metrics.
-        state.extracted_metrics.apply_enforcement(&enforcement);
-        enforcement.apply_with_outcomes(&mut state.managed_envelope);
-
-        if event_active {
-            state.remove_event();
-            debug_assert!(state.envelope().is_empty());
-        }
-
+        // Update cached rate limits with the freshly computed ones.
         if !limits.is_empty() {
-            self.inner
-                .addrs
-                .project_cache
-                .send(UpdateRateLimits::new(scoping.project_key, limits));
+            self.inner.addrs.project_cache.send(UpdateRateLimits::new(
+                state.managed_envelope.scoping().project_key,
+                limits,
+            ));
         }
 
         Ok(())
@@ -1429,21 +1454,27 @@ impl EnvelopeProcessorService {
             return Ok(());
         }
 
+        // If spans were already extracted for an event, we rely on span processing to extract metrics.
+        let extract_spans = !state.spans_extracted
+            && state.project_info.config.features.produces_spans()
+            && utils::sample(global.options.span_extraction_sample_rate.unwrap_or(1.0));
+
         let metrics = crate::metrics_extraction::event::extract_metrics(
             event,
-            state.spans_extracted,
             combined_config,
+            sampling_decision,
+            state.project_id,
             self.inner
                 .config
                 .aggregator_config_for(MetricNamespace::Spans)
                 .aggregator
                 .max_tag_value_length,
-            global.options.span_extraction_sample_rate,
+            extract_spans,
         );
 
         state
             .extracted_metrics
-            .extend_project_metrics(metrics, Some(sampling_decision));
+            .extend(metrics, Some(sampling_decision));
 
         if !state.project_info.has_feature(Feature::DiscardTransaction) {
             let transaction_from_dsc = state
@@ -1457,6 +1488,7 @@ impl EnvelopeProcessorService {
                 generic_config: Some(combined_config),
                 transaction_from_dsc,
                 sampling_decision,
+                target_project_id: state.project_id,
                 has_profile: profile_id.is_some(),
             };
 
@@ -1684,10 +1716,6 @@ impl EnvelopeProcessorService {
         let profile_id = profile::filter(state);
         profile::transfer_id(state, profile_id);
 
-        if_processing!(self.inner.config, {
-            attachment::create_placeholders(state);
-        });
-
         event::finalize(state, &self.inner.config)?;
         self.normalize_event(state)?;
 
@@ -1882,6 +1910,7 @@ impl EnvelopeProcessorService {
         mut managed_envelope: ManagedEnvelope,
         project_id: ProjectId,
         project_info: Arc<ProjectInfo>,
+        rate_limits: RateLimits,
         sampling_project_info: Option<Arc<ProjectInfo>>,
         reservoir_counters: Arc<Mutex<BTreeMap<RuleId, i64>>>,
     ) -> Result<ProcessingStateResult, ProcessingError> {
@@ -1906,6 +1935,7 @@ impl EnvelopeProcessorService {
                     managed_envelope,
                     project_id,
                     project_info,
+                    rate_limits,
                     sampling_project_info,
                     reservoir_counters,
                 );
@@ -1982,6 +2012,7 @@ impl EnvelopeProcessorService {
         let ProcessEnvelope {
             envelope: mut managed_envelope,
             project_info,
+            rate_limits,
             sampling_project_info,
             reservoir_counters,
         } = message;
@@ -2030,6 +2061,7 @@ impl EnvelopeProcessorService {
                     managed_envelope,
                     project_id,
                     project_info,
+                    rate_limits,
                     sampling_project_info,
                     reservoir_counters,
                 ) {
@@ -2547,7 +2579,7 @@ impl EnvelopeProcessorService {
         };
 
         let error_sample_rate = global_config.options.cardinality_limiter_error_sample_rate;
-        if !limits.exceeded_limits().is_empty() && sample(error_sample_rate) {
+        if !limits.exceeded_limits().is_empty() && utils::sample(error_sample_rate) {
             for limit in limits.exceeded_limits() {
                 relay_log::error!(
                     tags.organization_id = scoping.organization_id,
@@ -2904,6 +2936,70 @@ impl Service for EnvelopeProcessorService {
                     .await;
             }
         });
+    }
+}
+
+#[cfg(feature = "processing")]
+enum RateLimiter<'a> {
+    Cached,
+    Consistent(&'a RedisRateLimiter),
+}
+
+#[cfg(feature = "processing")]
+impl<'a> RateLimiter<'a> {
+    fn enforce<G>(
+        &self,
+        global_config: &GlobalConfig,
+        state: &mut ProcessEnvelopeState<G>,
+    ) -> Result<RateLimits, ProcessingError> {
+        if state.envelope().is_empty() && !state.has_event() {
+            return Ok(RateLimits::default());
+        }
+
+        let quotas = CombinedQuotas::new(global_config, state.project_info.get_quotas());
+        if quotas.is_empty() {
+            return Ok(RateLimits::default());
+        }
+
+        let event_category = state.event_category();
+
+        // When invoking the rate limiter, capture if the event item has been rate limited to also
+        // remove it from the processing state eventually.
+        let mut envelope_limiter =
+            EnvelopeLimiter::new(CheckLimits::All, |item_scope, quantity| match self {
+                RateLimiter::Cached => Ok(state.rate_limits.check_with_quotas(quotas, item_scope)),
+                RateLimiter::Consistent(rl) => Ok::<_, ProcessingError>(
+                    rl.is_rate_limited(quotas, item_scope, quantity, false)?,
+                ),
+            });
+
+        // Tell the envelope limiter about the event, since it has been removed from the Envelope at
+        // this stage in processing.
+        if let Some(category) = event_category {
+            envelope_limiter.assume_event(category);
+        }
+
+        let scoping = state.managed_envelope.scoping();
+        let (enforcement, limits) = metric!(timer(RelayTimers::EventProcessingRateLimiting), {
+            envelope_limiter.compute(state.managed_envelope.envelope_mut(), &scoping)?
+        });
+        let event_active = enforcement.is_event_active();
+
+        // Use the same rate limits as used for the envelope on the metrics.
+        // Those rate limits should not be checked for expiry or similar to ensure a consistent
+        // limiting of envelope items and metrics.
+        state
+            .extracted_metrics
+            .apply_enforcement(&enforcement, matches!(self, Self::Consistent(_)));
+        enforcement.apply_with_outcomes(&mut state.managed_envelope);
+
+        if event_active {
+            state.remove_event();
+            debug_assert!(state.envelope().is_empty());
+            debug_assert!(!state.has_event());
+        }
+
+        Ok(limits)
     }
 }
 
@@ -3464,6 +3560,7 @@ mod tests {
         let message = ProcessEnvelope {
             envelope,
             project_info: Arc::new(project_info),
+            rate_limits: Default::default(),
             sampling_project_info: None,
             reservoir_counters: ReservoirCounters::default(),
         };
@@ -3526,6 +3623,7 @@ mod tests {
         let process_message = ProcessEnvelope {
             envelope: managed_envelope,
             project_info: Arc::new(ProjectInfo::default()),
+            rate_limits: Default::default(),
             sampling_project_info: None,
             reservoir_counters: ReservoirCounters::default(),
         };