+++ relay-server/src/statsd.rs
@@ -181,6 +181,15 @@ pub enum RelayHistograms {
     /// Number of envelopes in the backpressure buffer between the envelope buffer
     /// and the project cache.
     BufferBackpressureEnvelopesCount,
+    /// The amount of bytes in the item payloads of an envelope pushed to the envelope buffer.
+    ///
+    /// This is not quite the same as the actual size of a serialized envelope, because it ignores
+    /// the envelope header and item headers.
+    BufferEnvelopeBodySize,
+    /// Size of a serialized envelope pushed to the envelope buffer (sampled).
+    BufferEnvelopeSize,
+    /// Size of a compressed envelope pushed to the envelope buffer (sampled).
+    BufferEnvelopeSizeCompressed,
     /// The number of batches emitted per partition.
     BatchesPerPartition,
     /// The number of buckets in a batch emitted.
@@ -309,6 +318,9 @@ impl HistogramMetric for RelayHistograms {
             RelayHistograms::BufferBackpressureEnvelopesCount => {
                 "buffer.backpressure_envelopes_count"
             }
+            RelayHistograms::BufferEnvelopeBodySize => "buffer.envelope_body_size",
+            RelayHistograms::BufferEnvelopeSize => "buffer.envelope_size",
+            RelayHistograms::BufferEnvelopeSizeCompressed => "buffer.envelope_size.compressed",
             RelayHistograms::ProjectStatePending => "project_state.pending",
             RelayHistograms::ProjectStateAttempts => "project_state.attempts",
             RelayHistograms::ProjectStateRequestBatchSize => "project_state.request.batch_size",
@@ -538,6 +550,10 @@ pub enum RelayTimers {
     BufferPop,
     /// Timing in milliseconds for the time it takes for the buffer to drain its envelopes.
     BufferDrain,
+    /// Timing in milliseconds for the time it takes for the envelopes to be serialized.
+    BufferEnvelopesSerialization,
+    /// Timing in milliseconds for the time it takes for the envelopes to be compressed (sampled).
+    BufferEnvelopeCompression,
 }
 
 impl TimerMetric for RelayTimers {
@@ -586,6 +602,8 @@ impl TimerMetric for RelayTimers {
             RelayTimers::BufferPeek => "buffer.peek.duration",
             RelayTimers::BufferPop => "buffer.pop.duration",
             RelayTimers::BufferDrain => "buffer.drain.duration",
+            RelayTimers::BufferEnvelopesSerialization => "buffer.envelopes_serialization",
+            RelayTimers::BufferEnvelopeCompression => "buffer.envelopes_compression",
         }
     }
 }